Mid-testing: removed PC printing type casting

+++ sim/main.c
@@ -216,7 +216,7 @@ main(int nargs, char *args[])
 	/*	In the non-LIBSF version, we use fprintf to write to console	*/
 	fprintf(stderr, "\n");
 	fprintf(stderr, "[ID=%d of %d][PC=0x" UHVLONGFMT "][%.1EV, %.1EMHz] ",
-		E->cp->NODE_ID, E->nnodes, (unsigned long)E->cp->PC,
+		E->cp->NODE_ID, E->nnodes, E->cp->PC,
 		E->cp->VDD, (1/E->cp->CYCLETIME)/1E6);
 
 	while (1)
@@ -236,7 +236,7 @@ main(int nargs, char *args[])
 				sf_riscv_parse();
 			}
 			fprintf(stderr, "[ID=%d of %d][PC=0x" UHVLONGFMT "][%.1EV, %.1EMHz] ",
-				E->cp->NODE_ID, E->nnodes, (unsigned long)E->cp->PC,
+				E->cp->NODE_ID, E->nnodes, E->cp->PC,
 				E->cp->VDD, (1/E->cp->CYCLETIME)/1E6);
 
 			/*
